fix: improve packages compatibility + better testing (#3233)

+++ scripts/customLinks.js
@@ -10,8 +10,10 @@ const path = require('path');
   const locationDir = './testapps/ngx/node_modules/@tolgee';
   const location = path.join(locationDir, 'ngx');
 
-  fs.rmSync(location, { recursive: true, force: true });
-  fs.symlinkSync(path.relative(locationDir, target), location);
+  if (fs.existsSync(location)) {
+    fs.rmSync(location, { recursive: true, force: true });
+    fs.symlinkSync(path.relative(locationDir, target), location);
+  }
 }
 
 {
@@ -19,6 +21,8 @@ const path = require('path');
   const locationDir = './testapps/svelte/node_modules/@tolgee';
   const location = path.join(locationDir, 'svelte');
 
-  fs.rmSync(location, { recursive: true, force: true });
-  fs.symlinkSync(path.relative(locationDir, target), location);
+  if (fs.existsSync(location)) {
+    fs.rmSync(location, { recursive: true, force: true });
+    fs.symlinkSync(path.relative(locationDir, target), location);
+  }
 }
